+++ packages/js-sdk/src/index.ts
@@ -1,2 +1,3 @@
 export { Snowball } from './Snowball'
+export type { MakeAuthOptions } from './Snowball'
 export { SnowballChain } from '@snowballtools/utils'